Move function definitions to src folder
