+++ main.cpp
@@ -1,4 +1,4 @@
-#include "include/graph.h"
+#include "graph.h"
 
 #include <iostream>
